requirements.txt
streamlit
python-dotenv
boto3
pypdf
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-pinecone
pinecone
ipykernel